PMM-10552 Drop dead code
