--- service/k8sclient/common/types.go
@@ -1,213 +1,0 @@
-// dbaas-controller
-// Copyright (C) 2020 Percona LLC
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-// Package common contains common API Schema definitions for the pxc and psmdb API groups
-package common
-
-// VolumeSpec backup storage volume specs.
-type VolumeSpec struct {
-	// EmptyDir to use as data volume for mysql. EmptyDir represents a temporary
-	// directory that shares a pod's lifetime.
-	EmptyDir *EmptyDirVolumeSource `json:"emptyDir,omitempty"`
-
-	// HostPath to use as data volume for mysql. HostPath represents a
-	// pre-existing file or directory on the host machine that is directly
-	// exposed to the container.
-	HostPath *HostPathVolumeSource `json:"hostPath,omitempty"`
-
-	// PersistentVolumeClaim to specify PVC spec for the volume for mysql data.
-	// It has the highest level of precedence, followed by HostPath and
-	// EmptyDir. And represents the PVC specification.
-	PersistentVolumeClaim *PersistentVolumeClaimSpec `json:"persistentVolumeClaim,omitempty"`
-}
-
-// PersistentVolumeClaimSpec describes the common attributes of storage devices
-// and allows a Source for provider-specific attributes.
-type PersistentVolumeClaimSpec struct {
-	StorageClassName string `json:"storageClassName,omitempty"`
-	// Resources represents the minimum resources the volume should have.
-	// More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#resources
-	Resources ResourceRequirements `json:"resources,omitempty"`
-}
-
-// ResourceRequirements describes the compute resource requirements.
-type ResourceRequirements struct {
-	// Limits describes the maximum amount of compute resources allowed.
-	// More info: https://kubernetes.io/docs/concepts/configuration/manage-compute-resources-container/
-	Limits ResourceList `json:"limits,omitempty"`
-	// Requests describes the minimum amount of compute resources required.
-	// If Requests is omitted for a container, it defaults to Limits if that is explicitly specified,
-	// otherwise to an implementation-defined value.
-	// More info: https://kubernetes.io/docs/concepts/configuration/manage-compute-resources-container/
-	Requests ResourceList `json:"requests,omitempty"`
-}
-
-// ResourceName is the name identifying various resources in a ResourceList.
-type ResourceName string
-
-// Resource names must be not more than 63 characters, consisting of upper- or lower-case alphanumeric characters,
-// with the -, _, and . characters allowed anywhere, except the first or last character.
-// The default convention, matching that for annotations, is to use lower-case names, with dashes, rather than
-// camel case, separating compound words.
-// Fully-qualified resource typenames are constructed from a DNS-style subdomain, followed by a slash `/` and a name.
-const (
-	// CPU, in cores. (500m = .5 cores).
-	ResourceCPU ResourceName = "cpu"
-	// Memory, in bytes. (500Gi = 500GiB = 500 * 1024 * 1024 * 1024).
-	ResourceMemory ResourceName = "memory"
-	// Volume size, in bytes (e,g. 5Gi = 5GiB = 5 * 1024 * 1024 * 1024).
-	ResourceStorage ResourceName = "storage"
-	// Local ephemeral storage, in bytes. (500Gi = 500GiB = 500 * 1024 * 1024 * 1024)
-	// The resource name for ResourceEphemeralStorage is alpha and it can change across releases.
-	ResourceEphemeralStorage ResourceName = "ephemeral-storage"
-)
-
-// ResourceList is a set of (resource name, quantity) pairs.
-type ResourceList map[ResourceName]string
-
-// ResourcesList POD's resources list.
-type ResourcesList struct {
-	Memory           string `json:"memory,omitempty"`
-	CPU              string `json:"cpu,omitempty"`
-	EphemeralStorage string `json:"ephemeral-storage,omitempty"`
-}
-
-// PodResources represents the POD resources limits.
-type PodResources struct {
-	Requests *ResourcesList `json:"requests,omitempty"`
-	Limits   *ResourcesList `json:"limits,omitempty"`
-}
-
-// PullPolicy describes a policy for if/when to pull a container image.
-type PullPolicy string
-
-const (
-	// PullAlways means that kubelet always attempts to pull the latest image. Container will fail If the pull fails.
-	PullAlways PullPolicy = "Always"
-	// PullNever means that kubelet never pulls an image, but only uses a local image. Container will fail if the image isn't present.
-	PullNever PullPolicy = "Never"
-	// PullIfNotPresent means that kubelet pulls if the image isn't present on disk. Container will fail if the image isn't present and the pull fails.
-	PullIfNotPresent PullPolicy = "IfNotPresent"
-)
-
-// PodDisruptionBudgetSpec POD disruption budget specs.
-type PodDisruptionBudgetSpec struct {
-	MinAvailable   *int `json:"minAvailable,omitempty"`
-	MaxUnavailable *int `json:"maxUnavailable,omitempty"`
-}
-
-// ServiceType represents the expose type to open the cluster to the world.
-type ServiceType string
-
-const (
-	// ServiceTypeClusterIP means a service will only be accessible inside the
-	// cluster, via the cluster IP.
-	ServiceTypeClusterIP ServiceType = "ClusterIP"
-
-	// ServiceTypeNodePort means a service will be exposed on one port of
-	// every node, in addition to 'ClusterIP' type.
-	ServiceTypeNodePort ServiceType = "NodePort"
-
-	// ServiceTypeLoadBalancer means a service will be exposed via an
-	// external load balancer (if the cloud provider supports it), in addition
-	// to 'NodePort' type.
-	ServiceTypeLoadBalancer ServiceType = "LoadBalancer"
-
-	// ServiceTypeExternalName means a service consists of only a reference to
-	// an external name that kubedns or equivalent will return as a CNAME
-	// record, with no exposing or proxying of any pods involved.
-	ServiceTypeExternalName ServiceType = "ExternalName"
-)
-
-// NodeFileSystemSummary holds a summary of Node's filesystem.
-type NodeFileSystemSummary struct {
-	UsedBytes uint64 `json:"usedBytes,omitempty"`
-}
-
-// NodeSummaryNode holds information about Node inside Node's summary.
-type NodeSummaryNode struct {
-	FileSystem NodeFileSystemSummary `json:"fs,omitempty"`
-}
-
-// NodeSummary holds summary of the Node.
-// One gets this by requesting Kubernetes API endpoint:
-// /v1/nodes/<node-name>/proxy/stats/summary.
-type NodeSummary struct {
-	Node NodeSummaryNode `json:"node,omitempty"`
-}
-
-// LocalObjectReference contains enough information to let you locate the
-// referenced object inside the same namespace.
-type LocalObjectReference struct {
-	// Name of the referent.
-	// More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names
-	Name string `json:"name,omitempty"`
-}
-
-// SecretKeySelector selects a key of a Secret.
-type SecretKeySelector struct {
-	// The name of the secret in the pod's namespace to select from.
-	LocalObjectReference `json:",inline"`
-	// The key of the secret to select from.  Must be a valid secret key.
-	Key string `json:"key"`
-	// Specify whether the Secret or its key must be defined
-	Optional *bool `json:"optional,omitempty"`
-}
-
-// UpgradeOptions holds configuration options to handle automatic upgrades.
-type UpgradeOptions struct {
-	VersionServiceEndpoint string `json:"versionServiceEndpoint,omitempty"`
-	Apply                  string `json:"apply,omitempty"`
-	Schedule               string `json:"schedule,omitempty"`
-}
-
-// AppState application state.
-type AppState string
-
-const (
-	// AppStateUnknown unknown application state.
-	AppStateUnknown AppState = "unknown"
-	// AppStateInit initializing application state.
-	AppStateInit AppState = "initializing"
-	// AppStateReady ready application state.
-	AppStateReady AppState = "ready"
-	// AppStateError error application state.
-	AppStateError AppState = "error"
-	// AppStatePaused paused application state. Available only in the operator >= 1.9.0.
-	AppStatePaused AppState = "paused"
-	// AppStateStopping means cluster is stopping. Available only in the operator >= 1.9.0.
-	AppStateStopping AppState = "stopping"
-)
-
-// DatabaseCluster is an abstraction of database cluster.
-type DatabaseCluster interface {
-	// SetDatabaseImage sets database image to appropriate image field.
-	SetDatabaseImage(name string)
-	// GetImage returns image of database software used.
-	DatabaseImage() string
-	// GetName returns name of the cluster.
-	GetName() string
-	// GetCRDName returns name of Custom Resource Definition -> cluster's kind.
-	CRDName() string
-	// DatabaseContainerNames returns container names that actually run the database.
-	DatabaseContainerNames() []string
-	// DatabasePodLabels return list of labels to get pods where database is running.
-	DatabasePodLabels() []string
-	// State get's clusters state as a string.
-	State() AppState
-	// Pause returns bool indicating if cluster should be paused.
-	Pause() bool
-}
